# Workaround to get shape key normals faster