@scurest 2024-01-26
Something to think about for the future... I'm pretty sure this is how getting shape key tangents should work too (I strongly suspect the current code for getting morph tangents is junk).