@julienduroure 2024-01-31
I closed the PR in blender repo ( https://projects.blender.org/blender/blender-addons/pulls/105134 )
to be sure that all discussion happen here, on upstream repo